Reformatting files for understandability
