+++ pug/pug.py
@@ -1,6 +1,6 @@
 """Commands for generating teams in pugs."""
 import random
-from typing import Optional, Dict, List, Set
+from typing import Optional, Dict, List
 
 import nextcord
 from nextcord.ext import commands
@@ -8,10 +8,6 @@ from nextcord.ext import commands
 from constants import BOT_COLOR
 from database import (
     BotCollection,
-    get_server,
-    add_med_immune_player,
-    remove_med_immune_player,
-    clear_med_immunity_by_guild,
 )
 from pug import (
     CategorySelect,
@@ -20,13 +16,119 @@ from pug import (
     MoveView,
     Player,
     PugCategory,
-    BooleanView,
 )
-from pug.setup import PugSetupCog
+
 from registration import RegistrationSettings
 from util import is_setup, is_runner
 
 category_db = BotCollection("guilds", "categories")
+
+
+async def get_player_dict(
+    next_pug: nextcord.VoiceChannel, add_up: nextcord.VoiceChannel
+) -> Dict[str, List[Player]]:
+    """Return a dict of players in two voice channels.
+
+    Args:
+        next_pug (nextcord.VoiceChannel): A voice channel to get players from.
+        add_up (nextcord.VoiceChannel): A voice channel to get players from.
+
+    Returns:
+        dict: A dict of players in the voice channels.
+    """
+    players: Dict[str, list[Player]] = {"next_pug": [], "add_up": []}
+    for member in next_pug.members:
+        player = Player(discord=member.id)
+        players["next_pug"].append(player)
+    for member in add_up.members:
+        player = Player(discord=member.id)
+        players["add_up"].append(player)
+    return players
+
+
+async def generate_random_teams(
+    players: Dict[str, List[Player]], team_size: int
+) -> Dict[str, list[Player]]:
+    """Generate random teams for a pug.
+
+    Args:
+        players (list): A list of all players in the VC
+        team_size (int): The amount of players per team.
+
+    Returns:
+        teams: The generated teams.
+    """
+    random.shuffle(players["next_pug"])
+    random.shuffle(players["add_up"])
+    all_players: List[Player] = players["next_pug"] + players["add_up"]
+
+    red_team: list[Player] = []
+    blu_team: list[Player] = []
+
+    while len(red_team) < team_size and len(blu_team) < team_size:
+        red_team.append(all_players.pop(0))
+        blu_team.append(all_players.pop(0))
+
+    teams = {"red": red_team, "blu": blu_team}
+    return teams
+
+
+async def generate_balanced_teams(
+    players: Dict[str, List[Player]],
+    team_size,
+    reg_settings: RegistrationSettings,
+) -> Dict[str, List[Player]]:
+    """Generate balanced teams for a pug.
+
+    Args:
+        players (dict): A dictionary of all players in the VCs
+        team_size (int): The amount of players per team.
+        reg_settings (RegistrationSettings): The registration settings for the server.
+
+    Returns:
+        teams: The generated teams.
+    """
+    if reg_settings.mode == "" or reg_settings.gamemode == "":
+        teams = await generate_random_teams(players, team_size)
+        return teams
+
+    gamemode: str
+    if reg_settings.gamemode == "sixes":
+        gamemode = "sixes"
+    elif reg_settings.gamemode == "highlander":
+        gamemode = "hl"
+
+    random.shuffle(players["next_pug"])
+    random.shuffle(players["add_up"])
+    players["next_pug"].sort(
+        key=lambda x: 10
+        if x.division[gamemode][reg_settings.mode] == -1
+        else x.division[gamemode][reg_settings.mode],
+        reverse=False,
+    )
+    players["add_up"].sort(
+        key=lambda x: 10
+        if x.division[gamemode][reg_settings.mode] == -1
+        else x.division[gamemode][reg_settings.mode],
+        reverse=False,
+    )
+    all_players = players["next_pug"] + players["add_up"]
+
+    red_team: list[Player] = []
+    blu_team: list[Player] = []
+    count = 0
+
+    while len(red_team) < team_size and len(blu_team) < team_size:
+        if count % 2 == 0:
+            red_team.append(all_players.pop(0))
+            blu_team.append(all_players.pop(0))
+        else:
+            blu_team.append(all_players.pop(0))
+            red_team.append(all_players.pop(0))
+        count += 1
+
+    teams = {"red": red_team, "blu": blu_team}
+    return teams
 
 
 class PugRunningCog(commands.Cog):
@@ -35,7 +137,14 @@ class PugRunningCog(commands.Cog):
     def __init__(self, bot: nextcord.Client):
         self.bot = bot
 
-    @PugSetupCog.pug.subcommand(  # pylint: disable=no-member
+    @nextcord.slash_command()
+    async def pug(self, interaction: nextcord.Interaction):
+        """
+        This is the main slash command that will be the prefix of all commands below.
+        This will never get called since it has subcommands.
+        """
+
+    @pug.subcommand(  # pylint: disable=no-member
         name="genteams", description="Generate teams for a pug."
     )
     @is_setup()
@@ -144,8 +253,8 @@ class PugRunningCog(commands.Cog):
         elif reg_settings.gamemode == "highlander":
             gamemode = "hl"
 
-        teams = await self.generate_balanced_teams(
-            await self.get_player_list(next_pug, add_up), team_size, reg_settings
+        teams = await generate_balanced_teams(
+            await get_player_dict(next_pug, add_up), team_size, reg_settings
         )
 
         while True:
@@ -243,124 +352,19 @@ class PugRunningCog(commands.Cog):
 
             if team_generation_view.action == "random":
                 pug_embed.clear_fields()
-                teams = await self.generate_random_teams(
-                    await self.get_player_list(next_pug, add_up), team_size
+                teams = await generate_random_teams(
+                    await get_player_dict(next_pug, add_up), team_size
                 )
 
             if team_generation_view.action == "balanced":
                 pug_embed.clear_fields()
-                teams = await self.generate_balanced_teams(
-                    await self.get_player_list(next_pug, add_up),
+                teams = await generate_balanced_teams(
+                    await get_player_dict(next_pug, add_up),
                     team_size,
                     reg_settings,
                 )
 
-    async def get_player_list(
-        self, next_pug: nextcord.VoiceChannel, add_up: nextcord.VoiceChannel
-    ) -> Dict[str, List[Player]]:
-        """Return a list of players in a voice channel.
-
-        Args:
-            channel (nextcord.VoiceChannel): The voice channel to get players from.
-
-        Returns:
-            list: A list of players in the voice channel.
-        """
-        players: Dict[str, list[Player]] = {"next_pug": [], "add_up": []}
-        for member in next_pug.members:
-            player = Player(discord=member.id)
-            players["next_pug"].append(player)
-        for member in add_up.members:
-            player = Player(discord=member.id)
-            players["add_up"].append(player)
-        return players
-
-    async def generate_random_teams(
-        self, players: Dict[str, List[Player]], team_size: int
-    ) -> Dict[str, list[Player]]:
-        """Generate random teams for a pug.
-
-        Args:
-            players (list): A list of all players in the VC
-            team_size (int): The amount of players per team.
-            reg_settings (RegistrationSettings): The registration settings for the server.
-
-        Returns:
-            teams: The generated teams.
-        """
-        random.shuffle(players["next_pug"])
-        random.shuffle(players["add_up"])
-        all_players: List[Player] = players["next_pug"] + players["add_up"]
-
-        red_team: list[Player] = []
-        blu_team: list[Player] = []
-
-        while len(red_team) < team_size and len(blu_team) < team_size:
-            red_team.append(all_players.pop(0))
-            blu_team.append(all_players.pop(0))
-
-        teams = {"red": red_team, "blu": blu_team}
-        return teams
-
-    async def generate_balanced_teams(
-        self,
-        players: Dict[str, List[Player]],
-        team_size,
-        reg_settings: RegistrationSettings,
-    ) -> Dict[str, List[Player]]:
-        """Generate balanced teams for a pug.
-
-        Args:
-            players (list): A list of all players in the VC
-            team_size (int): The amount of players per team.
-            reg_settings (RegistrationSettings): The registration settings for the server.
-
-        Returns:
-            teams: The generated teams.
-        """
-        if reg_settings.mode == "" or reg_settings.gamemode == "":
-            teams = await self.generate_random_teams(players, team_size)
-            return teams
-
-        gamemode: str
-        if reg_settings.gamemode == "sixes":
-            gamemode = "sixes"
-        elif reg_settings.gamemode == "highlander":
-            gamemode = "hl"
-
-        random.shuffle(players["next_pug"])
-        random.shuffle(players["add_up"])
-        players["next_pug"].sort(
-            key=lambda x: 10
-            if x.division[gamemode][reg_settings.mode] == -1
-            else x.division[gamemode][reg_settings.mode],
-            reverse=False,
-        )
-        players["add_up"].sort(
-            key=lambda x: 10
-            if x.division[gamemode][reg_settings.mode] == -1
-            else x.division[gamemode][reg_settings.mode],
-            reverse=False,
-        )
-        all_players = players["next_pug"] + players["add_up"]
-
-        red_team: list[Player] = []
-        blu_team: list[Player] = []
-        count = 0
-
-        while len(red_team) < team_size and len(blu_team) < team_size:
-            if count % 2 == 0:
-                red_team.append(all_players.pop(0))
-                blu_team.append(all_players.pop(0))
-            else:
-                blu_team.append(all_players.pop(0))
-                red_team.append(all_players.pop(0))
-            count += 1
-
-        teams = {"red": red_team, "blu": blu_team}
-        return teams
-
-    @PugSetupCog.pug.subcommand(  # pylint: disable=no-member
+    @pug.subcommand(  # pylint: disable=no-member
         name="move", description="Moves players after a pug is done."
     )
     @is_setup()
@@ -527,208 +531,3 @@ class PugRunningCog(commands.Cog):
             return
 
         await interaction.delete_original_message(delay=10)
-
-    @PugSetupCog.medic.subcommand(  # pylint: disable=no-member
-        name="roll",
-        description="Randomly selects a non-immune player to play medic.",
-    )
-    @is_setup()
-    @is_runner()
-    async def roll_medic(
-        self,
-        interaction: nextcord.Interaction,
-        team_size: Optional[int] = nextcord.SlashOption(
-            name="team_size",
-            description="The size of the teams, default 6",
-            required=False,
-            default=6,
-        ),
-    ):
-        await interaction.response.defer()
-
-        # Get a list of pug categories
-        try:
-            result = await category_db.find_item({"_id": interaction.guild.id})
-            categories = result["categories"]
-            if len(categories) == 0:
-                raise LookupError
-        except LookupError:
-            await interaction.send(
-                "There are no pug categories setup for this server.\nPlease run /pug category add to add a pug category."
-            )
-            return
-
-        select_view = CategorySelect()
-
-        for name, category in categories.items():
-            disabled: bool = False
-            color = nextcord.ButtonStyle.gray
-            add_up_channel: nextcord.VoiceChannel = interaction.guild.get_channel(
-                category["add_up"]
-            )
-            red_team_channel: nextcord.VoiceChannel = interaction.guild.get_channel(
-                category["red_team"]
-            )
-            blu_team_channel: nextcord.VoiceChannel = interaction.guild.get_channel(
-                category["blu_team"]
-            )
-            next_pug_channel: nextcord.VoiceChannel = interaction.guild.get_channel(
-                category["next_pug"]
-            )
-            if (
-                len(add_up_channel.members) + len(next_pug_channel.members)
-            ) < team_size * 2:
-                disabled = True
-                name += " (Not enough players)"
-            if len(red_team_channel.members) > 0 or len(blu_team_channel.members) > 0:
-                disabled = True
-                name += " (Pug in progress)"
-            if (
-                interaction.user.voice is not None
-                and interaction.user.voice.channel
-                == (next_pug_channel or add_up_channel)
-            ):
-                color = nextcord.ButtonStyle.green
-
-            button = CategoryButton(name=name, color=color, disabled=disabled)
-            select_view.add_item(button)
-
-        med_embed = nextcord.Embed(
-            title="Roll Medic",
-            color=BOT_COLOR,
-            description="Select the category you would like to roll medic for.",
-        )
-        await interaction.send(embed=med_embed, view=select_view)
-        embed_status = await select_view.wait()
-        if embed_status or select_view.name == "cancel":
-            return
-
-        chosen_category: PugCategory = PugCategory(
-            select_view.name, categories[select_view.name]
-        )
-
-        add_up: nextcord.VoiceChannel = interaction.guild.get_channel(
-            chosen_category.add_up
-        )
-        next_pug: nextcord.VoiceChannel = interaction.guild.get_channel(
-            chosen_category.next_pug
-        )
-
-        server = get_server(interaction.guild.id)
-        immune_players: Set[int] = set(server["immune"])
-
-        players = await self.get_player_list(next_pug, add_up)
-
-        immune_chosen = False  # True in the very rare case that all players in next pug/add up are already immune
-        medic: Player | None = None
-        # All players in the next pug must be picked from this channel including med
-        if len(players["next_pug"]) >= team_size * 2:
-            random.shuffle(players["next_pug"])
-            for player in players["next_pug"]:
-                if player.discord not in immune_players:
-                    medic = player
-                    break
-            if not medic:
-                medic = players["next_pug"][0]
-                immune_chosen = True
-        else:
-            random.shuffle(players["add_up"])
-            for player in players["add_up"]:
-                if player.discord not in immune_players:
-                    medic = player
-                    break
-            if not medic:
-                medic = players["add_up"][0]
-                immune_chosen = True
-
-        med_embed.description = f"Rolled <@{medic.discord}> as medic."
-
-        if not immune_chosen:
-            med_embed.description = "Give player medic roll immunity?"
-
-            boolean_view = BooleanView()
-
-            await interaction.edit_original_message(embed=med_embed, view=boolean_view)
-            status = await boolean_view.wait()
-
-            med_embed.clear_fields()
-            if status or not boolean_view.action:
-                med_embed.description = "Player not given med roll immunity"
-            else:
-                med_embed.description = "Player given med roll immunity"
-                add_med_immune_player(interaction.guild.id, medic.discord)
-
-        await interaction.edit_original_message(embed=med_embed, view=None)
-
-    @PugSetupCog.medic.subcommand(  # pylint: disable=no-member
-        name="immune", description="Manually set a player as immune to medic roll."
-    )
-    @is_setup()
-    @is_runner()
-    async def set_med_immune(
-        self,
-        interaction: nextcord.Interaction,
-        user: nextcord.User = nextcord.SlashOption(
-            name="discord", description="The Discord user to make immune", required=True
-        ),
-    ):
-        await interaction.response.defer()
-        add_med_immune_player(interaction.guild.id, user.id)
-
-        med_embed = nextcord.Embed(
-            title="Set User Immune to Medic Roll",
-            color=BOT_COLOR,
-            description=f"<@{user.id}> is now immune to medic roll.",
-        )
-        await interaction.send(embed=med_embed)
-
-    @PugSetupCog.medic.subcommand(  # pylint: disable=no-member
-        name="remove",
-        description="Manually remove a player from being immune to medic roll.",
-    )
-    @is_setup()
-    @is_runner()
-    async def remove_med_immune(
-        self,
-        interaction: nextcord.Interaction,
-        user: nextcord.User = nextcord.SlashOption(
-            name="discord",
-            description="The Discord user to remove immunity from",
-            required=True,
-        ),
-    ):
-        await interaction.response.defer()
-        remove_med_immune_player(interaction.guild.id, user.id)
-
-        med_embed = nextcord.Embed(
-            title="Removed User Immune to Medic Roll",
-            color=BOT_COLOR,
-            description=f"<@{user.id}> is no longer immune to medic roll.",
-        )
-        await interaction.send(embed=med_embed)
-
-    @PugSetupCog.medic.subcommand(  # pylint: disable=no-member
-        name="remove_all",
-        description="Manually set a player as immune to medic roll.",
-    )
-    @is_setup()
-    @is_runner()
-    async def remove_all_med_immune(self, interaction: nextcord.Interaction):
-        await interaction.response.defer()
-        clear_med_immunity_by_guild(interaction.guild.id)
-
-        med_embed = nextcord.Embed(
-            title="Remove Immunity From All Players",
-            color=BOT_COLOR,
-            description="Are you certain you want to clear medic roll immunity from all players?",
-        )
-        boolean_view = BooleanView()
-        await interaction.send(embed=med_embed, view=boolean_view)
-        view_status = await boolean_view.wait()
-
-        if view_status or not boolean_view.action:
-            med_embed.description = "Did not remove medic roll immunity."
-        else:
-            med_embed.description = "Removed medic roll immunity from all players."
-
-        await interaction.edit_original_message(embed=med_embed, view=None)
